detect event next, back to load ajax. Done Tests screen

+++ OnlineAptitudeExam/Areas/Admin/Controllers/TestsController.cs
@@ -3,11 +3,8 @@ using OnlineAptitudeExam.Utils;
 using PagedList;
 using System;
 using System.Data.Entity;
-using System.Data.Entity.Migrations;
 using System.Linq;
-using System.Web;
 using System.Web.Mvc;
-using System.Xml.Linq;
 
 namespace OnlineAptitudeExam.Areas.Admin.Controllers
 {
@@ -16,13 +13,15 @@ namespace OnlineAptitudeExam.Areas.Admin.Controllers
         private readonly OnlineAptitudeExamEntities db = new OnlineAptitudeExamEntities();
 
         // GET: Admin/Tests
-        // [Authentication(true)]
-        public ActionResult Index()
+        [Authentication(true)]
+        public ActionResult Index(bool isAjax = false)
         {
+            ViewBag.isAjax = isAjax;
             return View();
         }
 
         // GET: Admin/Tests/GetData
+        [Authentication(true)]
         public ActionResult GetData(string sort, string order, string filter, string search, int? page)
         {
             ViewBag.CurrentSort = sort;
@@ -40,24 +39,25 @@ namespace OnlineAptitudeExam.Areas.Admin.Controllers
                 switch (sort)
                 {
                     case SortHelper.NAME:
-                        tests = SortHelper.isAsc(order) ? tests.OrderBy(s => s.name) : tests.OrderByDescending(s => s.name);
+                        tests = SortHelper.IsAsc(order) ? tests.OrderBy(s => s.name) : tests.OrderByDescending(s => s.name);
                         break;
                     case SortHelper.DATE:
-                        tests = SortHelper.isAsc(order) ? tests.OrderBy(s => s.created_date) : tests.OrderByDescending(s => s.created_date);
+                        tests = SortHelper.IsAsc(order) ? tests.OrderBy(s => s.created_date) : tests.OrderByDescending(s => s.created_date);
                         break;
                     default:
-                        tests = SortHelper.isAsc(order) ? tests.OrderBy(s => s.id) : tests.OrderByDescending(s => s.id);
+                        tests = SortHelper.IsAsc(order) ? tests.OrderBy(s => s.id) : tests.OrderByDescending(s => s.id);
                         break;
                 }
             }
             else tests = tests.OrderByDescending(s => s.id);
 
             int pageSize = 5;
-            int pageNumber = ((page == null || page < 1) ? 1 : (int)page);
+            int pageNumber = (page == null || page < 1) ? 1 : (int)page;
             return View(tests.ToPagedList(pageNumber, pageSize));
         }
 
         [HttpPost, ValidateInput(false)]
+        [Authentication(true)]
         public ActionResult Create(string name)
         {
             if (ModelState.IsValid)
@@ -85,6 +85,7 @@ namespace OnlineAptitudeExam.Areas.Admin.Controllers
         }
 
         [HttpPost, ValidateInput(false)]
+        [Authentication(true)]
         public ActionResult Update(int id, string name)
         {
             if (ModelState.IsValid)
@@ -94,6 +95,10 @@ namespace OnlineAptitudeExam.Areas.Admin.Controllers
                     return Json(Responses.Error("Name must be not null!"), JsonRequestBehavior.AllowGet);
                 }
                 var test = db.Tests.Find(id);
+                if (test == null)
+                {
+                    return Json(Responses.Error("Not found item!"), JsonRequestBehavior.AllowGet);
+                }
                 if (test.Exams.Count() != 0)
                 {
                     return Json(Responses.Error("This test is already taken by users. You cannot change it!"), JsonRequestBehavior.AllowGet);
@@ -114,15 +119,30 @@ namespace OnlineAptitudeExam.Areas.Admin.Controllers
         }
 
         [HttpPost]
+        [Authentication(true)]
         public ActionResult ToggleStatus(int id)
         {
             if (ModelState.IsValid)
             {
                 var test = db.Tests.Find(id);
+                if (test == null)
+                {
+                    return Json(Responses.Error("Not found item!"), JsonRequestBehavior.AllowGet);
+                }
+                if (test.Exams.Count() < Constants.TOTAL_QUESTION_IN_TEST)
+                {
+                    if (test.status != (byte?)Enums.Status.PRIVATE)
+                    {
+                        test.status = (byte?)Enums.Status.PRIVATE;
+                        db.Entry(test).State = EntityState.Modified;
+                        db.SaveChanges();
+                    }
+                    return Json(Responses.Error("This test does not have enough questions.Please create more questions.", new {status = test.status}), JsonRequestBehavior.AllowGet);
+                }
                 test.status = Enums.GetOpposite(test.status.Value);
                 db.Entry(test).State = EntityState.Modified;
                 db.SaveChanges();
-                return Json(Responses.Success(test, "Update success!"), JsonRequestBehavior.AllowGet);
+                return Json(Responses.Success(new { status = test.status }, "Update success!"), JsonRequestBehavior.AllowGet);
             }
             else
             {
@@ -131,16 +151,21 @@ namespace OnlineAptitudeExam.Areas.Admin.Controllers
         }
 
         [HttpPost]
+        [Authentication(true)]
         public ActionResult Delete(int id)
         {
             if (ModelState.IsValid)
             {
                 var test = db.Tests.Find(id);
+                if(test == null)
+                {
+                    return Json(Responses.Error("Not found item!"), JsonRequestBehavior.AllowGet);
+                }
                 if (test.Exams.Count() != 0)
                 {
                     return Json(Responses.Error("This test is already taken by users. You cannot change it!"), JsonRequestBehavior.AllowGet);
                 }
-                if(test.Questions.Count() != 0)
+                if (test.Questions.Count() != 0)
                 {
                     db.Questions.RemoveRange(test.Questions);
                 }
@@ -152,6 +177,14 @@ namespace OnlineAptitudeExam.Areas.Admin.Controllers
             {
                 return Json(Responses.Error("Something was wrong"), JsonRequestBehavior.AllowGet);
             }
+        }
+
+        [Authentication(true)]
+        public ActionResult Detail(int id, bool isAjax = false)
+        {
+            var test = db.Tests.Find(id);
+            ViewBag.IsAjax = isAjax;
+            return View(test);
         }
     }
